Move containers to the correct location
